* commit '1060068311253b25748178f410989dd090475e22 [formerly e511ee206b1cc3c704a37c4dd992f575ac991945]':
  Skip delete for query folders.
  Deny rename for "Shared with me" folder.
  Include "Shared with me" volume.
  Revert.
  Remove "Starred" volume.
  Define constant for "Shared with me" folder.
  Map volumes to root id.
  Display trashed files as hidden.
  Optimize list API call to only include fiels in response that are required.
  Disallow moving virtual directories.
  Allow new files in root.
  Display files of "My Drive" in root.
  Add starred list service.
  Disable for virtual containers.
  Disable for virtual containers.
  Fix lookup for virtual container to return root node.
  Add parent virtual containers.
  Rename test.
  Extract list services to allow any query.
  Update dependency.


Former-commit-id: 26db7eaf1dab7d34ae49aa738f2f63f90805ce4e

+++ googledrive/src/main/java/ch/cyberduck/core/googledrive/DriveDeleteFeature.java
@@ -34,6 +34,9 @@ public class DriveDeleteFeature implements Delete {
     @Override
     public void delete(final List<Path> files, final LoginCallback prompt, final Callback callback) throws BackgroundException {
         for(Path file : files) {
+            if(file.getType().contains(Path.Type.placeholder)) {
+                continue;
+            }
             callback.delete(file);
             try {
                 session.getClient().files().delete(new DriveFileidProvider(session).getFileid(file)).execute();
